Move useLocation calls closer to where they are used
